successfully setup redux store witj basic add and remove functionality and buttons for view, edit and delete operations
