+++ package.json
@@ -19,6 +19,7 @@
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.17.0",
     "react-scripts": "^5.0.1",
+    "uuid": "^9.0.1",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
