--- src/pages/InvoiceStore.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const InvoiceStore = () => {
-	return <div>InvoiceStore</div>
-}
-
-export default InvoiceStore
