+++ src/store/store.js
@@ -1,0 +1,10 @@
+import { configureStore } from '@reduxjs/toolkit'
+import invoiceReducer from './invoiceSlice'
+
+const store = configureStore({
+	reducer: {
+		invoice: invoiceReducer
+	}
+})
+
+export default store
